migrating interfaces to a central store
